pyproject.toml
[tool.poetry]
name = "repl_python3_anti-spy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
discord = "^1.7.3"
bs4 = "^0.0.1"
django-encrypt = "^0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
